Refactor deployment actions to reduce number of files
